backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-genai==0.6.0
pydantic==2.9.2
